+++ src/App.jsx
@@ -188,7 +188,7 @@ function App() {
       <LanguageProvider>
         <CurrencyProvider>
           <ThemeProvider>
-            <Router>
+            <Router future={{ v7_startTransition: true }}>
               <AppContent />
             </Router>
           </ThemeProvider>